refactor: abstract core marketplace listing logic to vehiclefeed component
