--- components/Footer.js
@@ -1,11 +1,0 @@
-import styles from './Footer.module.css'
-
-export default function Footer() {
-  return (
-    <>
-      <footer className={styles.footer}>
-        <img src="/logo.png" alt="Netlify Logo" className={styles.logo} />
-      </footer>
-    </>
-  )
-}
