--- components/Footer.module.css
@@ -1,13 +1,0 @@
-.footer {
-  width: 100%;
-  height: 100px;
-  border-top: 1px solid #eaeaea;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.logo {
-  height: 3em;
-  margin: 5px;
-}
