--- components/VehicleListing.module.css
@@ -1,48 +1,0 @@
-.listing-item {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  border: 1px solid #ddd;
-  padding: 10px;
-  border-radius: 5px;
-  background-color: #f9f9f9;
-  text-align: center;
-  margin: 10px;
-  flex: 0 0 calc((100% - 30px) / 3);
-  max-width: calc((100% - 30px) / 3);
-  border: 1px solid #ddd;
-}
-
-.listing-image {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 5px;
-  overflow: hidden;
-  width: 100%;
-  object-fit: cover;
-}
-
-.listing-details {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.listing-title {
-  font-size: 16px;
-  font-weight: bold;
-  margin: 5px 0;
-}
-
-.listing-description {
-  font-size: 14px;
-  color: #666;
-  margin: 5px 0;
-}
-
-.listing-price {
-  font-size: 14px;
-  font-weight: bold;
-  color: #333;
-}
