+++ styles/home.css
@@ -22,12 +22,12 @@ body {
     justify-content: space-between;
     width: 100%; 
     height: 75px;
-    box-sizing: border-box; /* Include padding in width calculation */
-    position: fixed; /* Make the navbar fixed */
-    top: 0; /* Fix it to the top of the viewport */
-    left: 0; /* Ensure it starts at the left edge */
-    z-index: 1000; /* Ensure it stays above other content */
-    border-bottom: 1px solid #ddd; /* Optional: Add a bottom border for better visual separation */
+    box-sizing: border-box;
+    position: fixed; 
+    top: 0;
+    left: 0;
+    z-index: 1000;
+    border-bottom: 1px solid #ddd;
 }
